Generate BMS Wiring Diagram
This was generated on https://www.orionbms.com/tools/wiring.php?cell_count=95&fuse_position%5B%5D=20&fuse_position%5B%5D=39&fuse_position%5B%5D=58&fuse_position%5B%5D=77.

There are 19 cells in each row of the accumulator, so halfway through the middle group, the cells are all tied to the cell that is a multiple of 19.  In the left group, they are all listed as N/A because they have their own separate ground.